requirements.txt
Pillow==10.0.1
pyee==11.0.0
python-ffmpeg==2.0.4
typing_extensions==4.8.0
